+++ docs/conf.py
@@ -12,6 +12,7 @@ rst_epilog = """
 .. _mypy: http://www.mypy-lang.org
 .. _pre-commit: https://pre-commit.com
 .. _pytest: https://docs.pytest.org/en/latest/
+.. _Strimzi: https://strimzi.io
 .. _tox: https://tox.readthedocs.io/en/latest/
 """